move ERC20Api to net
